DABOM-341 refactor: Reader 공통화
